bot/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
